third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-angle-space-list-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <angle>+
PASS Animating a custom property of type <angle>+ with a single keyframe
PASS Animating a custom property of type <angle>+ with additivity
PASS Animating a custom property of type <angle>+ with a single keyframe and additivity
FAIL Animating a custom property of type <angle>+ with iterationComposite assert_equals: expected "250deg 275deg" but got "50deg 75deg"
PASS Animating a custom property of type <angle>+ with different lengths is discrete
Harness: the test ran to completion.
